requirements.txt
# Python 依赖列表
# 用于 yuntucv_web 项目的代理服务器
# 兼容 Python 3.6+ 版本

# Web 框架 - 使用兼容较低Python版本的版本
Flask==2.2.5
flask-cors==3.0.10

# HTTP 请求库 - 使用兼容版本
requests==2.28.2

# 图像处理 - 使用兼容版本
Pillow==9.5.0
# EXR 支持需要 OpenEXR
OpenEXR==1.3.9

# 其他标准库依赖（Python 内置，无需安装）
# - json
# - os
# - base64
# - io

# 可选：用于生产环境部署
# gunicorn==20.1.0
# gevent==22.10.2